Add files via upload
Here are all the required files in order for this code to run. Please message me if there are others however this should be all the files. The majority of the files are for the phylogenetic tree to be created since it was difficult for me to run a pairwise sequence alignment since the fasta files for the spike protein would not import in the first half without error messages.